Updated skepu::external.
SkePU's external call now has support for specifying whether containers should
be read or written to during an external call. A container can be listed in
both the read and write specifiers.

In the MPI backend, containers has a new function called init. This function
will initialize a container to a specified size and with an optional default
value.